fix: write header when flushing (FIR-47197)
**Background**

[FIR-47197](https://packboard.atlassian.net/browse/FIR-47197): Write iceberg manifest file(s)

As part of writing Iceberg tables, we need to write key-value metadata
to avro files.  The last several commits from ClickHouse make the
metadata mutable, but segfault when closing a DataFileWriter without
having written any rows to it yet, because the OutputStream in the
BinaryEncoder is not created yet.

**Summary**

Fix the segfault in DataFileWriterBase::close()/flush() by flushing the
avro header, which forces the BinaryEncoder to initialize.

**Testing**

The PackDB unit test SchemaInference.AvroSchemaReaderInferStruct stops
segfaulting with this change.

Wrote some throwaway code that is able to write a file with zero rows,
which crashed before this change, and after this change successfully
wrote a file with a correct looking schema and 0 rows.